webapp styling
Hey guys, here is a colour scheme for the webapp! I also did a small be of css styling for it so let me know what you think! I want to add the dublin bikes logo to the header and some other things, but feel free to give me suggestions of other things to add!